+++ data/duck.txt
@@ -1,0 +1,4 @@
+T;asd;0
+E;asd;10/10/0010 2100;1
+T;asd;0
+E;asd;10/10/0010 1010;0
