+++ src/main/java/Commands/AddCommand.java
@@ -1,3 +1,9 @@
+package Commands;
+import Models.Todo;
+import Duck.TaskList;
+import Duck.Storage;
+import Duck.UI;
+
 public class AddCommand extends Commands{
     private Todo item;
     public AddCommand(Todo item) {
