+++ src/main/java/Commands/DeleteCommand.java
@@ -1,4 +1,8 @@
-import java.io.IOException;
+package Commands;
+import Models.Todo;
+import Duck.TaskList;
+import Duck.Storage;
+import Duck.UI;
 
 public class DeleteCommand extends Commands{
     private int index;
