+++ src/main/java/Commands/ErrorCommand.java
@@ -1,3 +1,8 @@
+package Commands;
+import Models.Todo;
+import Duck.TaskList;
+import Duck.Storage;
+
 public class ErrorCommand extends Commands{
     public ErrorCommand(){
