+++ src/main/java/Commands/ExitCommand.java
@@ -1,3 +1,8 @@
+package Commands;
+import Models.Todo;
+import Duck.TaskList;
+import Duck.Storage;
+
 public class ExitCommand extends Commands{
     public ExitCommand(){
     }
