+++ src/main/java/Duck/Duck.java
@@ -1,3 +1,6 @@
+package Duck;
+import Models.Todo;
+import Commands.Commands;
 import java.io.*;
 import java.nio.file.Path;
 import java.text.ParseException;
