+++ src/main/java/Duck/Parser.java
@@ -1,4 +1,19 @@
+package Duck;
+import Commands.Commands;
+import Commands.ExitCommand;
+import Commands.ListCommand;
+import Commands.AddCommand;
+import Commands.MarkCommand;
+import Commands.UnmarkCommand;
+import Commands.DeleteCommand;
+import Commands.ErrorCommand;
+import Models.Todo;
+import Models.Event;
+import Models.Deadline;
+import Quackceptions.UnallowedCharacterException;
+
 import java.text.ParseException;
+
 
 public class Parser {
