+++ src/main/java/Models/Todo.java
@@ -1,3 +1,4 @@
+package Models;
 import java.io.BufferedWriter;
 import java.io.FileWriter;
 import java.io.IOException;
