+++ src/main/java/Quackceptions/UnallowedCharacterException.java
@@ -1,3 +1,4 @@
+package Quackceptions;
 public class UnallowedCharacterException extends Exception{
     public UnallowedCharacterException(String message) {
         super(message);
